:art: #3089【微信支付】新增代金券详情接口返回类增加商户单据号字段

+++ weixin-java-pay/src/main/java/com/github/binarywang/wxpay/bean/marketing/FavorStocksGetResult.java
@@ -112,7 +112,7 @@ public class FavorStocksGetResult implements Serializable {
    * 示例值：100
    */
   @SerializedName("distributed_coupons")
-  private Integer distributedCoupons;
+  private Long distributedCoupons;
 
   /**
    * 是否无资金流
